# openapi.yaml
openapi: 3.0.3
info:
  title: Calendar Aggregator API
  description: |
    A powerful calendar aggregation API that combines multiple iCal feeds into a unified calendar service.
    
    Features:
    - **Combine multiple calendars** from different sources (Google Calendar, Outlook, Apple Calendar, etc.)
    - **Validate and test** calendar URLs before adding them
    - **Fetch events** from multiple calendars concurrently
    - **Monitor sync status** and handle errors gracefully
    
    Built with Next.js 15, TypeScript, and modern web technologies.
  version: 1.0.0
  contact:
    name: Calendar Aggregator Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
paths:
  /calendars:
    get:
      summary: List all calendars
      description: Retrieve all configured calendar sources with their current sync status
      operationId: getCalendars
      tags:
        - Calendars
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/CalendarSource'
                  count:
                    type: integer
                    description: Total number of calendars
              example:
                calendars:
                  - id: 1
                    name: "Work Calendar"
                    url: "https://calendar.google.com/calendar/ical/work@example.com/public/basic.ics"
                    color: "#3b82f6"
                    enabled: true
                    syncStatus: "success"
                    createdAt: "2024-01-01T00:00:00.000Z"
                    lastSyncAt: "2024-01-01T12:00:00.000Z"
                count: 1
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: Add a new calendar
      description: Create a new calendar source with URL validation and connection testing
      operationId: createCalendar
      tags:
        - Calendars
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCalendarRequest'
            example:
              name: "Personal Calendar"
              url: "webcal://calendar.google.com/calendar/ical/personal@example.com/public/basic.ics"
              color: "#ef4444"
              enabled: true
      responses:
        '201':
          description: Calendar created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalendarSource'
        '400':
          description: Invalid request or calendar URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                invalid_url:
                  summary: Invalid URL format
                  value:
                    error: "Invalid calendar URL"
                    details: "Invalid URL format"
                connection_failed:
                  summary: Connection test failed
                  value:
                    error: "Invalid calendar URL"
                    details: "Unable to connect to server"
                    warnings: ["Content-Type is not text/calendar"]
        '409':
          description: Calendar URL already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Calendar URL already exists"
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
    CalendarSource:
      type: object
      required:
        - id
        - url
        - name
        - color
        - enabled
        - createdAt
        - syncStatus
      properties:
        id:
          type: integer
          description: Unique calendar identifier
        url:
          type: string
          format: uri
          description: iCal feed URL (webcal:// URLs are converted to https://)
        name:
          type: string
          description: Human-readable calendar name
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          description: Hex color code for calendar display
          example: "#3b82f6"
        enabled:
          type: boolean
          description: Whether calendar is active for syncing
        createdAt:
          type: string
          format: date-time
          description: Calendar creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last modification timestamp
        lastSyncAt:
          type: string
          format: date-time
          nullable: true
          description: Last successful sync timestamp
        syncStatus:
          type: string
          enum: [idle, syncing, error, success]
          description: Current sync status
        errorMessage:
          type: string
          nullable: true
          description: Last error message if sync failed
    CreateCalendarRequest:
      type: object
      required:
        - url
        - name
      properties:
        url:
          type: string
          format: uri
          description: iCal feed URL (supports webcal:// protocol)
          example: "https://calendar.google.com/calendar/ical/example@gmail.com/public/basic.ics"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Calendar display name
          example: "Work Calendar"
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          description: Hex color code for calendar display
          default: "#3b82f6"
        enabled:
          type: boolean
          description: Whether calendar should be enabled for syncing
          default: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Main error message
        details:
          type: string
          description: Detailed error information
        warnings:
          type: array
          items:
            type: string
          description: Additional warnings
  responses:
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
...